% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_edge_dataframe.R
\name{build_edge_dataframe}
\alias{build_edge_dataframe}
\title{The length of a string (in characters).}
\usage{
build_edge_dataframe(df, iter, disc.col.palette = NULL,
  missing.disc.col = "#888888")
}
\arguments{
\item{df}{dataframe containing the ACSSS study data}

\item{iter}{the year/location to build edges for}
}
\description{
The length of a string (in characters).
}
