% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_node_dataframe.R
\name{build_node_dataframe}
\alias{build_node_dataframe}
\title{The length of a string (in characters).}
\usage{
build_node_dataframe(df, iter, disc.col.palette = NULL,
  missing.disc.col = "#888888")
}
\arguments{
\item{df}{dataframe containing the ACSSS study data}

\item{iter}{the year/location to build nodes for}
}
\description{
The length of a string (in characters).
}
