% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophily_functions.R
\name{get_ei}
\alias{get_ei}
\title{Krackhardt and Stern E-I Index
Relative density of within-group connections compare to number of extra-group connections}
\usage{
get_ei(nodes, edges, column)
}
\arguments{
\item{nodes}{the dataframe containing the network nodes}

\item{edges}{the dataframe contianing the network edges}

\item{column}{the name of the column (as a string) to calculate homophily for}
}
\description{
number of ties in different groups - number of ties in same group/ number of total ties
}
\details{
Range
+ -1 is completely homophilious
+ 1 is completely heterophilious
}
