% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophily_functions.R
\name{get_hhi}
\alias{get_hhi}
\title{Herfindahl-Hirschman Index (HHI)
Sum of market shares of groups, where an increase in value indicates increasing
concentration of alters in one large group while a decrease in value suggests a
more equal distribution of alters among smaller groups.}
\usage{
get_hhi(nodes, edges, column)
}
\arguments{
\item{nodes}{the dataframe containing the network nodes}

\item{edges}{the dataframe contianing the network edges}

\item{column}{the name of the column (as a string) to calculate homophily for}
}
\description{
sum of the square(number of alters in given group/total alters) for all possible groups
}
\details{
Range
+ 0 is completely heterophilious
+ 10,000 represents perfect homophily
}
