% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentile.R
\name{get_percentile}
\alias{get_percentile}
\title{Given a distribution and a value, x, returns the percentile value
of x given the distirbution}
\usage{
get_percentile(dist, x)
}
\arguments{
\item{dist}{the distribution from which to compute the ecdf}

\item{x}{the value for which we want to know the percentile}
}
\description{
Given a distribution and a value, x, returns the percentile value
of x given the distirbution
}
