% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_acsss_data.R
\name{load_acsss_data}
\alias{load_acsss_data}
\title{The length of a string (in characters).}
\usage{
load_acsss_data(path)
}
\arguments{
\item{path}{path to the ACSSS study data file}
}
\description{
The length of a string (in characters).
}
